+++ pyproject.toml
@@ -11,10 +11,10 @@ docker = "^5.0.3"
 gym = "^0.23.1"
 pydantic = "^1.9.2"
 ray = {extras = ["rllib"], version = "^2.0.0"}
-tensorflow = "^2.10.0"
 fancyflags = "^1.1"
 absl-py = "^1.3.0"
 wandb = "^0.13.4"
+torch = "^1.13.0"
 
 [tool.poetry.dev-dependencies]
 ipython = "^8.4.0"
